requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
python-dotenv==1.0.1
requests==2.31.0
streamlit==1.32.0
faster-whisper
torch==2.2.0+cpu
streamlit-mic-recorder==0.0.5
langchain==0.2.8
langchain-community==0.2.6
langchain-groq==0.1.3
langchain-huggingface==1.0.1
faiss-cpu==1.7.4
sentence-transformers==2.5.1
numpy==1.26.4
scikit-learn==1.4.2
tavily-python==0.1.7
